What the hell happened here

+++ minilibft/ft_substr.c
@@ -6,16 +6,16 @@
 /*   By: moel-asr <moel-asr@student.1337.ma>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/10/09 14:51:39 by moel-asr          #+#    #+#             */
-/*   Updated: 2023/06/10 17:00:00 by moel-asr         ###   ########.fr       */
+/*   Updated: 2023/07/09 21:44:18 by moel-asr         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "minilibft.h"
 
-char	*ft_substr(char const *s, unsigned int start, size_t len)
+char	*ft_substr(char const *s, int start, int len)
 {
 	char	*sub;
-	size_t	i;
+	int		i;
 
 	if (!s)
 		return (NULL);
